src/requirements.txt
click==8.1.3
PyYAML==6.0
mistune==2.0.4